redesigned save, export and export dropdown menu bar
