+++ oar-dmp/src/app/app.component.html
@@ -5,6 +5,34 @@
 <div *ngIf="creds && readyDisplay">
     <div class="dmp-frm-row">
         <div class="dmp-col-7">
+                <div class="dmp_control_bar">
+                    <div class="grid_buttons" style="padding-top: 3px;">
+                        <!-- <button mat-button type="button" class="btn_reset" (click)="dmpButtonClick($event)">
+                        Reset
+                        </button> -->
+                        <div>
+                        <button mat-button type="button" id="btnSave" class="btn_draft" (click)="dmpButtonClick($event)" [disabled]="disableSaveBtn">
+                            Save
+                        </button>
+                        </div>
+                    
+                        <div>
+                        <button mat-button type="button" class="btn_draft" (click)="dmpButtonClick($event)">
+                            Export As:
+                        </button>
+                        </div>
+                        
+                        <div class="dmpExportSelect" >
+                        <select name="exportType" [(ngModel)]="exportType" (change)="setExportFormat()" >
+                            <option value="" disabled selected>Export As:</option>
+                            <option [value]="item.id" *ngFor="let item of exportFormats">
+                                {{item.format}}
+                            </option>
+                        </select>
+                        </div>
+                    </div>
+                </div>
+            
             <div id="dmp_panel">
                 <div class="dmp_header" >Data Management Plan</div>
