+++ oar-dmp/src/app/resource-options/resource-options.component.ts
@@ -8,13 +8,6 @@ import { Subscription } from 'rxjs';
 // In the child, we need to import the service "ResourcesService" file to be able to use it. 
 import { ResourcesService } from '../shared/resources.service';
 import { LoadResourcesService } from '../shared/load-resources.service';
-import { SubmitDmpService } from '../shared/submit-dmp.service';
-import { DropDownSelectService } from '../shared/drop-down-select.service';
-import { FormChangedService } from '../shared/form-changed.service';
-
-interface Messages {
-  [key: string]: any;
-}
 
 
 @Component({
@@ -28,7 +21,7 @@ export class ResourceOptionsComponent implements OnInit, AfterViewInit {
   softwareSubscription!: Subscription | null;
   databaseSubscription!: Subscription | null;
   websiteSubscription!: Subscription | null;
-  formChangedSubscription!: Subscription | null;
+  
 
   // to pass data from parent to this component we declare an input decorator along
   // with a property named data wich is of type string
@@ -37,10 +30,7 @@ export class ResourceOptionsComponent implements OnInit, AfterViewInit {
   // we inject shared service ResourcesService in the constructor.
   constructor(
     private sharedService:ResourcesService, 
-    private nistResources: LoadResourcesService,
-    private form_buttons: SubmitDmpService,
-    private dropDownService: DropDownSelectService,
-    private formChangedService: FormChangedService
+    private nistResources: LoadResourcesService    
     ) { 
 
   }
@@ -58,21 +48,7 @@ export class ResourceOptionsComponent implements OnInit, AfterViewInit {
   softwareSelection: string = "";
   dataSelection: string = "";
 
-  availableResources: any = {};
-
-  exportType: string = "";
-  exportFormats = [
-    {
-      id: "1",
-      format: 'PDF'
-    },
-    {
-      id: "2",
-      format: 'Markdown'
-    }
-  ];
-
-  disableSaveBtn:boolean = false;
+  availableResources: any = {};   
 
   ngOnInit(): void {
     // this.message = this.sharedService.getMessage()
@@ -80,9 +56,8 @@ export class ResourceOptionsComponent implements OnInit, AfterViewInit {
     this.softwareSubscribe();
     this.databaseSubscribe();
     this.websiteSubscribe();
-    this.storageSubscribe();
     this.availableResources = this.nistResources.getAllResources();
-    this.saveButtonSubscribe();
+    
     // console.log(this.availableResources);
   }
   ngAfterViewInit(): void {
@@ -91,32 +66,9 @@ export class ResourceOptionsComponent implements OnInit, AfterViewInit {
     // this.dom.horizontalDomAdjust("resource_options", "dmp_hdr")
   }
 
-  setExportFormat(){
-    let dataFormat = this.dropDownService.getDropDownText(this.exportType, this.exportFormats)[0].format;
-    this.form_buttons.setexportFormat(dataFormat);
-    this.form_buttons.exportFormatSubject$.next(dataFormat)
-  }
+  
 
-  dmpButtonClick(e:any){
-    //send message to dmp form component indicating which button has been pressed
-    //the e event captures the text of the button so we pass tat to the form to
-    //indicate the course of action. The options should be 'Reset', 'Save' and 'Publish'
-    this.form_buttons.setButtonMessage(e.currentTarget.innerText);
-    this.form_buttons.buttonSubject$.next(e.currentTarget.innerText);
-    
-  }
-
-  //subscribe to a particular subject
-  saveButtonSubscribe() {
-    if (!this.formChangedSubscription) {
-      //subscribe if not already subscribed
-      this.formChangedSubscription = this.formChangedService.disableSaveBtn$.subscribe({
-        next: (message) => {
-          this.disableSaveBtn = message;          
-        }
-      });
-    }
-  }
+  
 
   //subscribe to a particular subject
   storageSubscribe() {
